+++ busy_beaver/tasks/__init__.py
@@ -1,1 +1,1 @@
-from .tasks import post_github_summary_to_slack  # noqa
+from .github_stats import post_github_summary_to_slack  # noqa
