+++ tests/tasks/github_stats_test.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 
 from busy_beaver.models import User
-from busy_beaver.tasks.github_stats import generate_summary
+from busy_beaver.tasks.github_stats.summary import generate_summary
 from busy_beaver.toolbox import utc_now_minus
 
 import pytest
